--- lib/features/chat/domain/repositories/user_repository.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_chat/features/chat/domain/entities/user_entity.dart';
-
-abstract class UserRepository {
-  Stream<List<UserEntity>> getUserByUserName(String username);
-}
