--- lib/features/chat/presentation/pages/chat_room_list_tile.dart
@@ -1,50 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_chat/services/database.dart';
-
-class ChatRoomListTile extends StatefulWidget {
-  final String lastMessage, chatRoomId, myUsername;
-  ChatRoomListTile(this.lastMessage, this.chatRoomId, this.myUsername);
-
-  @override
-  _ChatRoomListTileState createState() => _ChatRoomListTileState();
-}
-
-class _ChatRoomListTileState extends State<ChatRoomListTile> {
-  String profilePicUrl = "", name = "", username = "";
-
-  getThisUserInfo() async {
-    username =
-        widget.chatRoomId.replaceAll(widget.myUsername, "").replaceAll("_", "");
-    QuerySnapshot querySnapshot = await DatabaseMethods().getUserInfo(username);
-    name = "${querySnapshot.docs[0]["name"]}";
-    profilePicUrl = "${querySnapshot.docs[0]["imgUrl"]}";
-    setState(() {});
-  }
-
-  @override
-  void initState() {
-    getThisUserInfo();
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        CircleAvatar(
-          backgroundImage: NetworkImage(profilePicUrl),
-        ),
-        SizedBox(
-          width: 12,
-        ),
-        Column(
-          children: [
-            Text(name),
-            Text(widget.lastMessage),
-          ],
-        )
-      ],
-    );
-  }
-}
