--- lib/services/database.dart
@@ -1,81 +1,0 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter_chat/helper_functions/shared_preferences_helper.dart';
-
-class DatabaseMethods {
-  Future addUserInfoToDB(
-      String userId, Map<String, dynamic> userInfoMap) async {
-    return await FirebaseFirestore.instance
-        .collection("users")
-        .doc(userId)
-        .set(userInfoMap);
-  }
-
-  Future<Stream<QuerySnapshot>> getUserByUserName(String username) async {
-    return FirebaseFirestore.instance
-        .collection("users")
-        .where("username", isEqualTo: username)
-        .snapshots();
-  }
-
-  Future addMessage(String chatRoomId, String messageId,
-      Map<String, dynamic> messageInfoMap) async {
-    return FirebaseFirestore.instance
-        .collection("chatrooms")
-        .doc(chatRoomId)
-        .collection("chats")
-        .doc(messageId)
-        .set(messageInfoMap);
-  }
-
-  updateLastMessageSend(
-      String chatRoomId, Map<String, dynamic> lastMessageInfoMap) {
-    return FirebaseFirestore.instance
-        .collection("chatrooms")
-        .doc(chatRoomId)
-        .update(lastMessageInfoMap);
-  }
-
-  createChatRoom(
-      String chatRoomId, Map<String, dynamic> chatRoomInfoMap) async {
-    final snapShot = await FirebaseFirestore.instance
-        .collection("chatrooms")
-        .doc(chatRoomId)
-        .get();
-
-    if (snapShot.exists) {
-      // chatroom already exist
-      return true;
-    } else {
-      // chatroom does not exists
-      return FirebaseFirestore.instance
-          .collection("chatrooms")
-          .doc(chatRoomId)
-          .set(chatRoomInfoMap);
-    }
-  }
-
-  Future<Stream<QuerySnapshot>> getChatRoomMessages(chatRoomId) async {
-    return FirebaseFirestore.instance
-        .collection("chatrooms")
-        .doc(chatRoomId)
-        .collection("chats")
-        .orderBy("ts", descending: true)
-        .snapshots();
-  }
-
-  Future<Stream<QuerySnapshot>> getChatRooms() async {
-    String? myUsername = await SharedPreferenceHelper().getUserName();
-    return FirebaseFirestore.instance
-        .collection("chatrooms")
-        .orderBy("lastMessageSendTs", descending: true)
-        .where("users", arrayContains: myUsername)
-        .snapshots();
-  }
-
-  Future<QuerySnapshot> getUserInfo(String username) async {
-    return await FirebaseFirestore.instance
-        .collection("users")
-        .where("username", isEqualTo: username)
-        .get();
-  }
-}
