+++ lib/src/common/widgets/error_widget.dart
