+++ lib/src/core/error/exception.dart
