+++ lib/src/domain/entities/user_entity.dart
