+++ lib/src/domain/repositories/user_repository.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_chat/src/domain/entities/user_entity.dart';
+
+abstract class UserRepository {
+  Stream<List<UserEntity>> getUserByUserName(String username);
+}
